[feature] #2 - rootVC -> navVC 전환
