+++ wanted_pre_onboarding/ViewController/WeatherDetail/WeatherDetailViewController.swift
@@ -1,0 +1,68 @@
+//
+//  WeatherDetailViewController.swift
+//  wanted_pre_onboarding
+//
+//  Created by BH on 2022/06/13.
+//
+
+import UIKit
+
+class WeatherDetailViewController: UIViewController {
+    
+    // MARK: - Properties
+
+    // MARK: - LifeCycle
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        configureUI()
+    }
+    
+    // MARK: - Methods
+    
+    private func configureUI() {
+        view.backgroundColor = .brown
+    }
+    
+    // MARK: - @objc
+    
+}
+
+// MARK: - Extensions
+
+
+
+// 전처리
+#if DEBUG
+
+import SwiftUI
+@available(iOS 13.0, *)
+
+// UIViewControllerRepresentable을 채택
+struct ViewControllerRepresentable: UIViewControllerRepresentable {
+    // update
+    // _ uiViewController: UIViewController로 지정
+    func updateUIViewController(_ uiViewController: UIViewController , context: Context) {
+        
+    }
+    // makeui
+    func makeUIViewController(context: Context) -> UIViewController {
+        // Preview를 보고자 하는 Viewcontroller 이름
+        // e.g.)
+        return WeatherViewController()
+    }
+}
+
+struct ViewController_Previews: PreviewProvider {
+    
+    @available(iOS 13.0, *)
+    static var previews: some View {
+        // UIViewControllerRepresentable에 지정된 이름.
+        ViewControllerRepresentable()
+        
+        // 테스트 해보고자 하는 기기
+            .previewDevice("iPhone 11")
+    }
+}
+#endif
