+++ wanted_pre_onboarding/ViewController/WeatherList/WeatherViewController.swift
@@ -97,40 +97,9 @@ extension WeatherViewController: UICollectionViewDataSource {
 
 extension WeatherViewController: UICollectionViewDelegate {
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        print("index: \(weatherOfCity[indexPath.row])")
-    }
-}
-
-// 전처리
-#if DEBUG
-
-import SwiftUI
-@available(iOS 13.0, *)
-
-// UIViewControllerRepresentable을 채택
-struct ViewControllerRepresentable: UIViewControllerRepresentable {
-    // update
-    // _ uiViewController: UIViewController로 지정
-    func updateUIViewController(_ uiViewController: UIViewController , context: Context) {
         
-    }
-    // makeui
-    func makeUIViewController(context: Context) -> UIViewController {
-        // Preview를 보고자 하는 Viewcontroller 이름
-        // e.g.)
-        return WeatherViewController()
-    }
-}
-
-struct ViewController_Previews: PreviewProvider {
-    
-    @available(iOS 13.0, *)
-    static var previews: some View {
-        // UIViewControllerRepresentable에 지정된 이름.
-        ViewControllerRepresentable()
+        let weatherDetailVC = WeatherDetailViewController()
         
-        // 테스트 해보고자 하는 기기
-            .previewDevice("iPhone 11")
+        navigationController?.pushViewController(weatherDetailVC, animated: true)
     }
 }
-#endif
